+++ docker/init-db/scripts/init_lemon.sql
@@ -9,3 +9,4 @@ CREATE TABLE IF NOT EXISTS lemonldap_accounts (
 INSERT INTO lemonldap_accounts (username, password, cn, mail, roles) VALUES ('admin', 'password', 'admin name', 'admin@example.com', 'admin,user');
 INSERT INTO lemonldap_accounts (username, password, cn, mail, roles) VALUES ('user1', 'password', 'user1 name', 'user1@example.com', 'user');
 INSERT INTO lemonldap_accounts (username, password, cn, mail, roles) VALUES ('user2', 'password', 'user2 name', 'user2@example.com', 'user');
+INSERT INTO lemonldap_accounts (username, password, cn, mail, roles) VALUES ('noaccount', 'password', 'noaccount name', 'noaccount@example.com', 'user');